pyproject.toml
[tool.poetry]
name = "medical-app"
version = "0.1.0"
description = ""
authors = ["SabirovaOlya <sabirovaolya44@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
celery = "^5.4.0"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
django-jazzmin = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
